Implmented document uploading and management
implmented the insert, select, delete, and update, queries for the Document management section.

Still need to change select so that you can set what application it is attached to but have to wait for the applications to be setup first.

also if needed will change it so that the select only selects rows that pertain to that user rather than just all rows.