finished the last query for program information management
finished up the delete query for program information management.
you can now full delete a program which deletes it along with any information of it in other tables.

you can also just remove access to it which removes all the rows pertaining to it from the Track table so that students are no longer able to access it.